refactor: move types from service to types.d.ts
