+++ src/app/resolvers/route.resolver.ts
@@ -1,18 +1,20 @@
-import { Injectable } from '@angular/core';
-import { RouterStateSnapshot, ActivatedRouteSnapshot } from '@angular/router';
-import { Observable } from 'rxjs';
+import {Injectable} from '@angular/core';
+import {RouterStateSnapshot, ActivatedRouteSnapshot} from '@angular/router';
+import {Observable} from 'rxjs';
 
-import { DataService, IWrappedSummaryData } from '../data.service';
+import {DataService} from '../data.service';
+import {IWrappedSummaryData} from "../dashboard/types";
 
 @Injectable({
-  providedIn: 'root'
+    providedIn: 'root'
 })
-export class RouteResolver  {
-  
-  constructor(private dataService: DataService) {}
+export class RouteResolver {
 
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<IWrappedSummaryData> {
-    console.log('>>> RouteResolver >>> ', route.params, state)
-    return this.dataService.requestSummary(route.params)
-  }
+    constructor(private dataService: DataService) {
+    }
+
+    resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<IWrappedSummaryData> {
+        console.log('>>> RouteResolver >>> ', route.params, state)
+        return this.dataService.requestSummary(route.params)
+    }
 }
